Clean up the shorties components into smaller chunks
